Remove unused dependencies and add check to tests (#4870)
