+++ CHANGELOG.md
@@ -81,6 +81,7 @@
 - Removed layer re-ordering, layer sorting, layer type selection from UI [\#4616](https://github.com/raster-foundry/raster-foundry/pull/4616)
 - Removed lingering database assocation of scenes and projects (in deference to scenes to layers to projects) [\#4764](https://github.com/raster-foundry/raster-foundry/pull/4764) [\#4867](https://github.com/raster-foundry/raster-foundry/pull/4867)
 - Removed tool categories and tags [\#4779](https://github.com/raster-foundry/raster-foundry/pull/4779)
+- Removed unused dependencies and cleaned up the build file [\#4870](https://github.com/raster-foundry/raster-foundry/pull/4870)
 
 ### Fixed
