--- app-backend/batch/src/test/resources/export/localJob.json
@@ -1,130 +1,0 @@
-{
-  "id": "dda6080f-f7ad-455d-b409-764dd8c57039",
-  "input": {
-    "projectId": "dda6080f-f7ad-455d-b409-764dd8c57036",
-    "resolution": 15,
-    "style": {
-      "layers": [
-        {
-          "sceneId": "8436f7e9-b7f7-4d4f-bda8-76b32c356cff",
-          "ingestLocation": "s3://test/",
-          "colorCorrections": {
-            "redBand": 0,
-            "blueBand": 2,
-            "greenBand": 1,
-            "bandClipping": {
-              "enabled": true,
-              "redMax": 9,
-              "greenMax": 9,
-              "blueMax": 9,
-              "redMin": null,
-              "greenMin": null,
-              "blueMin": null
-            },
-            "sigmoidalContrast": {
-              "enabled": true,
-              "alpha": 6.0,
-              "beta": 7.0
-            },
-            "tileClipping": {
-              "enabled": true,
-              "min": 8,
-              "max": 9
-            },
-            "gamma": {
-              "enabled": true,
-              "redGamma": 1.0,
-              "greenGamma": 2.0,
-              "blueGamma": 3.0
-            },
-            "equalize": {
-              "enabled": true
-            },
-            "saturation": {
-              "enabled": true,
-              "saturation": 1.0
-            },
-            "autoBalance": {
-              "enabled": false
-            }
-          },
-          "sceneType": "AVRO",
-          "footprint": {
-            "type":"MultiPolygon",
-            "coordinates":[
-              [
-                [
-                  [
-                    -78.91337999999999,
-                    -12.25582
-                  ],
-                  [
-                    -77.20025,
-                    -12.620750000000017
-                  ],
-                  [
-                    -76.82601,
-                    -10.880060000000018
-                  ],
-                  [
-                    -78.52844,
-                    -10.519079999999995
-                  ],
-                  [
-                    -78.91337999999999,
-                    -12.25582
-                  ]
-                ]
-              ]
-            ]
-          },
-          "singleBandOptions": {},
-          "isSingleBand": false
-        }
-      ],
-      "mask": {
-        "type": "MultiPolygon",
-        "coordinates": [
-          [
-            [
-              [
-                -78.91337999999999,
-                -12.25582
-              ],
-              [
-                -77.20025,
-                -12.620750000000017
-              ],
-              [
-                -76.82601,
-                -10.880060000000018
-              ],
-              [
-                -78.52844,
-                -10.519079999999995
-              ],
-              [
-                -78.91337999999999,
-                -12.25582
-              ]
-            ]
-          ]
-        ]
-      }
-    }
-  },
-  "output": {
-    "source": "s3://test/",
-    "render": {
-      "operation": "id",
-      "bands": [
-        1,
-        2,
-        3
-      ]
-    },
-    "crop": false,
-    "rasterSize": 256,
-    "crs": "epsg:32654"
-  }
-}
