--- app-backend/batch/src/test/resources/export/testFileJob.json
@@ -1,41 +1,0 @@
-{
-    "id":"dda6080f-f7ad-455d-b409-764dd8c57039",
-    "input":{
-        "projectId":"6de62bb2-5cd9-41c9-b5f8-d2902f8b8725",
-        "resolution": 8,
-        "mask": {"type":"MultiPolygon","coordinates":[[[[-67.50000000407866,2.8113711946224087],[-67.50000000407866,8.407168167437906],[-61.875000001492985,8.407168167437906],[-61.875000001492985,2.8113711946224087],[-67.50000000407866,2.8113711946224087]]]]},
-        "layers":[
-            {
-                "layerId":"ca10532c-0432-4bb2-9d89-9ab8ba4b7e51",
-                "ingestLocation":"s3://rasterfoundry-staging-catalogs-us-east-1/layers",
-                "colorCorrections": {
-                    "redBand": 4,
-                    "greenBand": 3,
-                    "blueBand": 2,
-                    "redMax": 9, "greenMax": 9, "blueMax": 9,
-                    "redMin": null, "greenMin": null, "blueMin": null,
-                    "redGamma": 0.5,
-                    "greenGamma": 0.5,
-                    "blueGamma": 0.5,
-                    "alpha": null,
-                    "min": 4500,
-                    "max": 17000,
-                    "equalize": false,
-                    "beta": null,
-                    "brightness": null,
-                    "contrast": null,
-                    "autoBalance": false
-                }
-            }
-        ]
-    },
-    "output":{
-        "source":"file:///tmp",
-        "crop":false,
-        "render":{
-            "operation":"id",
-            "bands":[0, 1, 2]
-        },
-        "rasterSize": 1024
-    }
-}
