--- app-backend/batch/src/test/resources/ingest/awsJob.json
@@ -1,46 +1,0 @@
-{
-    "id": "dda6080f-f7ad-455d-b409-764dd8c57039",
-    "layers": [{
-        "id": "8436f7e9-b7f7-4d4f-bda8-76b32c356dff",
-        "output": {
-            "uri": "s3://rasterfoundry-staging-catalogs-us-east-1/test",
-            "crs": "epsg:3857",
-            "cellType": "uint16raw",
-            "histogramBuckets": 512,
-            "tileSize": 256,
-            "resampleMethod": "NearestNeighbor",
-            "keyIndexMethod": "ZCurveKeyIndexMethod",
-            "pyramid": true,
-            "native": true
-        },
-        "sources": [{
-            "uri": "http://landsat-pds.s3.amazonaws.com/L8/107/035/LC81070352015218LGN00/LC81070352015218LGN00_B4.TIF",
-            "extent": [138.8339238,  34.9569460, 141.4502449,  37.1094577],
-            "extentCrs": "epsg:32654",
-            "cellSize": {"width": 37.151447651062888, "height": -37.151447651062888},
-            "crs": "epsg:32654",
-            "bandMaps": [{"source": 1, "target": {"name": "1", "index": 1}}]
-        }, {
-            "uri": "http://landsat-pds.s3.amazonaws.com/L8/107/035/LC81070352015218LGN00/LC81070352015218LGN00_B3.TIF",
-            "extent": [138.8339238,  34.9569460, 141.4502449,  37.1094577],
-            "extentCrs": "epsg:32654",
-            "cellSize": {"width": 37.151447651062888, "height": -37.151447651062888},
-            "crs": "epsg:32654",
-            "bandMaps": [{"source": 1, "target": {"name": "2", "index": 2}}]
-        }, {
-            "uri": "http://landsat-pds.s3.amazonaws.com/L8/107/035/LC81070352015218LGN00/LC81070352015218LGN00_B2.TIF",
-            "extent": [138.8339238,  34.9569460, 141.4502449,  37.1094577],
-            "extentCrs": "epsg:32654",
-            "cellSize": {"width": 37.151447651062888, "height": -37.151447651062888},
-            "crs": "epsg:32654",
-            "bandMaps": [{"source": 1, "target": {"name": "3", "index": 3}}]
-        }, {
-            "uri": "http://landsat-pds.s3.amazonaws.com/L8/107/035/LC81070352015218LGN00/LC81070352015218LGN00_B5.TIF",
-            "extent": [138.8339238,  34.9569460, 141.4502449,  37.1094577],
-            "extentCrs": "epsg:32654",
-            "cellSize": {"width": 37.151447651062888, "height": -37.151447651062888},
-            "crs": "epsg:32654",
-            "bandMaps": [{"source": 1, "target": {"name": "4", "index": 4}}]
-        }]
-    }]
-}
