--- app-backend/batch/src/test/resources/ingest/localJob.json
@@ -1,46 +1,0 @@
-{
-    "id": "dda6080f-f7ad-455d-b409-764dd8c57039",
-    "layers": [{
-        "id": "8436f7e9-b7f7-4d4f-bda8-76b32c356cff",
-        "output": {
-            "uri": "file:///tmp/landsat-rgb",
-            "crs": "epsg:3857",
-            "cellType": "uint16raw",
-            "histogramBuckets": 256,
-            "tileSize": 256,
-            "resampleMethod": "NearestNeighbor",
-            "keyIndexMethod": "ZCurveKeyIndexMethod",
-            "pyramid": true,
-            "native": true
-        },
-        "sources": [{
-            "uri": "file:///opt/rf/test-resources/imagery/landsat-clipped-B4.tif",
-            "extent": [138.8339238,  34.9569460, 141.4502449,  37.1094577],
-            "extentCrs": "epsg:32654",
-            "cellSize": {"width": 37.151447651062888, "height": -37.151447651062888},
-            "crs": "epsg:32654",
-            "bandMaps": [{"source": 1, "target": {"name": "1", "index": 1}}]
-        }, {
-            "uri": "file:///opt/rf/test-resources/imagery/landsat-clipped-B3.tif",
-            "extent": [138.8339238,  34.9569460, 141.4502449,  37.1094577],
-            "extentCrs": "epsg:32654",
-            "cellSize": {"width": 37.151447651062888, "height": -37.151447651062888},
-            "crs": "epsg:32654",
-            "bandMaps": [{"source": 1, "target": {"name": "2", "index": 2}}]
-        }, {
-            "uri": "file:///opt/rf/test-resources/imagery/landsat-clipped-B2.tif",
-            "extent": [138.8339238,  34.9569460, 141.4502449,  37.1094577],
-            "extentCrs": "epsg:32654",
-            "cellSize": {"width": 37.151447651062888, "height": -37.151447651062888},
-            "crs": "epsg:32654",
-            "bandMaps": [{"source": 1, "target": {"name": "3", "index": 3}}]
-        }, {
-            "uri": "file:///opt/rf/test-resources/imagery/landsat-clipped-B5.tif",
-            "extent": [138.8339238,  34.9569460, 141.4502449,  37.1094577],
-            "extentCrs": "epsg:32654",
-            "cellSize": {"width": 37.151447651062888, "height": -37.151447651062888},
-            "crs": "epsg:32654",
-            "bandMaps": [{"source": 1, "target": {"name": "4", "index": 4}}]
-        }]
-    }]
-}
