--- app-backend/batch/src/test/resources/logback-test.xml
@@ -1,15 +1,0 @@
-<configuration>
-  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-    <encoder>
-      <pattern>[%thread] %-5level - %msg%n</pattern>
-    </encoder>
-  </appender>
-
-  <logger name="org.apache.http.wire" level="info" additivity="false">
-    <appender-ref ref="STDOUT" />
-  </logger>
-
-  <root level="debug">
-    <appender-ref ref="STDOUT" />
-  </root>
-</configuration>
