--- app-backend/batch/src/test/scala/com/azavea/rf/batch/BatchSpec.scala
@@ -1,11 +1,0 @@
-package com.rasterfoundry.batch
-
-trait BatchSpec {
-  def getJson(resource: String): String = {
-    val stream = getClass.getResourceAsStream(resource)
-    val lines = scala.io.Source.fromInputStream(stream).getLines
-    val json = lines.mkString(" ")
-    stream.close()
-    json
-  }
-}
