+++ app-backend/project/Dependencies.scala
@@ -2,12 +2,65 @@ import scala.util.Properties
 
 import sbt._
 
+object Version {
+  val akka = "2.5.11"
+  val akkaCirceJson = "1.22.0"
+  val akkaHttp = "10.0.11"
+  val akkaHttpCors = "0.2.2"
+  val akkaHttpExtensions = "0.4.15"
+  val akkaSlf4j = "2.4.11"
+  val apacheCommonsEmail = "1.5"
+  val auth0 = "1.5.0"
+  val awsBatchSdk = "1.11.196"
+  val awsStsSdk = "1.11.232"
+  val betterFiles = "3.4.0"
+  val caffeine = "2.3.5"
+  val cats = "1.4.0"
+  val catsEffect = "1.0.0"
+  val catsMeow = "0.2.0"
+  val chill = "0.9.2"
+  val circe = "0.11.1"
+  val circeOptics = "0.11.0"
+  val commonsIO = "2.5"
+  val dnsJava = "2.1.8"
+  val doobie = "0.6.0"
+  val dropbox = "3.0.9"
+  val elasticacheClient = "1.1.1"
+  val ficus = "1.4.0"
+  val geotools = "17.1"
+  val geotrellis = "2.2.0"
+  val geotrellisServer = "0.1.15"
+  val hadoop = "2.8.4"
+  val hikariCP = "3.2.0"
+  val http4s = "0.20.0-M6"
+  val json4s = "3.5.0"
+  val jts = "1.16.0"
+  val maml = "0.2.1"
+  val nimbusJose = "0.6.0"
+  val postgres = "42.1.1"
+  val postgis = "2.2.1"
+  val rollbar = "1.4.0"
+  val scaffeine = "2.0.0"
+  val scala = "2.11.12"
+  val scalaCheck = "1.14.0"
+  val scalacache = "0.27.0"
+  val scalaForklift = "0.3.1"
+  val scalaLogging = "3.9.0"
+  val scalaTest = "3.0.1"
+  val scalajHttp = "2.3.0"
+  val scapegoat = "1.3.7"
+  val scopt = "3.5.0"
+  val decline = "0.6.0"
+  val slickMigrationAPI = "0.4.0"
+  val spark = "2.2.0"
+  val sup = "0.2.0"
+}
+
 object Dependencies {
-  val akka = "com.typesafe.akka" %% "akka-actor" % Version.akka
   val akkaCirceJson = "de.heikoseeberger" %% "akka-http-circe" % Version.akkaCirceJson
   val akkaHttpCors = "ch.megard" %% "akka-http-cors" % Version.akkaHttpCors
   val akkaHttpExtensions = "com.lonelyplanet" %% "akka-http-extensions" % Version.akkaHttpExtensions
-  val akkaSlf4j = "com.typesafe.akka" %% "akka-slf4j" % Version.akkaSlf4j
+  val akkaSlf4j = "com.typesafe.akka" %% "akka-slf4j" % Version.akkaSlf4j % Runtime
   val akkahttp = "com.typesafe.akka" %% "akka-http" % Version.akkaHttp
   val akkastream = "com.typesafe.akka" %% "akka-stream" % Version.akka
   val akkatestkit = "com.typesafe.akka" %% "akka-http-testkit" % Version.akkaHttp
@@ -28,7 +81,6 @@ object Dependencies {
   val circeParser = "io.circe" %% "circe-parser" % Version.circe
   val circeTest = "io.circe" %% "circe-testing" % Version.circe % "test"
   val commonsIO = "commons-io" % "commons-io" % Version.commonsIO
-  val dnsJava = "dnsjava" % "dnsjava" % Version.dnsJava
   val doobieCore = "org.tpolecat" %% "doobie-core" % Version.doobie
   val doobieHikari = "org.tpolecat" %% "doobie-hikari" % Version.doobie
   val doobiePostgres = "org.tpolecat" %% "doobie-postgres" % Version.doobie
@@ -38,37 +90,37 @@ object Dependencies {
   val dropbox = "com.dropbox.core" % "dropbox-core-sdk" % Version.dropbox
   val elasticacheClient = "com.amazonaws" % "elasticache-java-cluster-client" % Version.elasticacheClient
   val ficus = "com.iheart" %% "ficus" % Version.ficus
-  val findbugAnnotations = "com.google.code.findbugs" % "annotations" % Version.findbugAnnotations % "compile"
-  val geotools = "org.geotools" % "gt-shapefile" % Version.geotools
   val geotrellisGeotools = "org.locationtech.geotrellis" %% "geotrellis-geotools" % Version.geotrellis
   val geotrellisRaster = "org.locationtech.geotrellis" %% "geotrellis-raster" % Version.geotrellis
   val geotrellisS3 = "org.locationtech.geotrellis" %% "geotrellis-s3" % Version.geotrellis
   val geotrellisServer = "com.azavea" %% "geotrellis-server-core" % Version.geotrellisServer
+
   val geotrellisServerOgc = "com.azavea" %% "geotrellis-server-ogc" % Version.geotrellisServer
   val geotrellisShapefile = "org.locationtech.geotrellis" %% "geotrellis-shapefile" % Version.geotrellis
   val geotrellisSpark = "org.locationtech.geotrellis" %% "geotrellis-spark" % Version.geotrellis
   val geotrellisUtil = "org.locationtech.geotrellis" %% "geotrellis-util" % Version.geotrellis
   val geotrellisVector = "org.locationtech.geotrellis" %% "geotrellis-vector" % Version.geotrellis
   val geotrellisVectorTestkit = "org.locationtech.geotrellis" %% "geotrellis-vector-testkit" % Version.geotrellis % "test"
-  val hadoopAws = "org.apache.hadoop" % "hadoop-aws" % Version.hadoop
   val hikariCP = "com.typesafe.slick" %% "slick-hikaricp" % Version.hikariCP
   val http4sBlaze = "org.http4s" %% "http4s-blaze-server" % Version.http4s
   val http4sBlazeClient = "org.http4s" %% "http4s-blaze-client" % Version.http4s
   val http4sCirce = "org.http4s" %% "http4s-circe" % Version.http4s
   val http4sDSL = "org.http4s" %% "http4s-dsl" % Version.http4s
   val http4sServer = "org.http4s" %% "http4s-server" % Version.http4s
-  val jts = "org.locationtech.jts" % "jts-core" % Version.jts
   val mamlJvm = "com.azavea" %% "maml-jvm" % Version.maml
   val mamlSpark = "com.azavea" %% "maml-spark" % Version.maml
   val nimbusJose = "com.guizmaii" %% "scala-nimbus-jose-jwt" % Version.nimbusJose
   val postgres = "org.postgresql" % "postgresql" % Version.postgres
+  val postgis = "net.postgis" % "postgis-jdbc" % Version.postgis
   val rollbar = "com.rollbar" % "rollbar-java" % Version.rollbar
   val scaffeine = "com.github.blemale" %% "scaffeine" % Version.scaffeine
+  val scalacacheCore = "com.github.cb372" %% "scalacache-core" % Version.scalacache
+  val scalacacheCats = "com.github.cb372" %% "scalacache-cats-effect" % Version.scalacache
+  val scalacacheCaffeine = "com.github.cb372" %% "scalacache-caffeine" % Version.scalacache
+  val scalacacheMemcached = "com.github.cb372" %% "scalacache-memcached" % Version.scalacache intransitive ()
   val scalaCheck = "org.scalacheck" %% "scalacheck" % Version.scalaCheck % "test"
-  val scalaCsv = "com.github.tototoshi" %% "scala-csv" % Version.scalaCsv
   val scalaLogging = "com.typesafe.scala-logging" %% "scala-logging" % Version.scalaLogging
   val scalaforklift = "com.liyaos" %% "scala-forklift-slick" % Version.scalaForklift
-  val scalajHttp = "org.scalaj" %% "scalaj-http" % Version.scalajHttp
   val scalatest = "org.scalatest" %% "scalatest" % Version.scalaTest % "test"
   val scopt = "com.github.scopt" %% "scopt" % Version.scopt
   val decline = "com.monovore" %% "decline" % Version.decline
