--- app-backend/project/Version.scala
@@ -1,53 +1,0 @@
-object Version {
-  val akka = "2.5.11"
-  val akkaCirceJson = "1.22.0"
-  val akkaHttp = "10.0.11"
-  val akkaHttpCors = "0.2.2"
-  val akkaHttpExtensions = "0.4.13"
-  val akkaSlf4j = "2.4.11"
-  val apacheCommonsEmail = "1.5"
-  val auth0 = "1.5.0"
-  val awsBatchSdk = "1.11.196"
-  val awsStsSdk = "1.11.232"
-  val betterFiles = "3.4.0"
-  val caffeine = "2.3.5"
-  val cats = "1.4.0"
-  val catsEffect = "1.0.0"
-  val catsMeow = "0.2.0"
-  val chill = "0.9.2"
-  val circe = "0.11.1"
-  val circeOptics = "0.11.0"
-  val commonsIO = "2.5"
-  val dnsJava = "2.1.8"
-  val doobie = "0.6.0"
-  val dropbox = "3.0.9"
-  val elasticacheClient = "1.1.1"
-  val ficus = "1.4.0"
-  val findbugAnnotations = "3.0.1u2"
-  val geotools = "17.1"
-  val geotrellis = "2.2.0"
-  val geotrellisServer = "0.1.15"
-  val hadoop = "2.8.4"
-  val hikariCP = "3.2.0"
-  val http4s = "0.20.0-M6"
-  val json4s = "3.5.0"
-  val jts = "1.16.0"
-  val maml = "0.2.1"
-  val nimbusJose = "0.6.0"
-  val postgres = "42.1.1"
-  val rollbar = "1.4.0"
-  val scaffeine = "2.0.0"
-  val scala = "2.11.12"
-  val scalaCheck = "1.14.0"
-  val scalaCsv = "1.3.5"
-  val scalaForklift = "0.3.1"
-  val scalaLogging = "3.9.0"
-  val scalaTest = "3.0.1"
-  val scalajHttp = "2.3.0"
-  val scapegoat = "1.3.7"
-  val scopt = "3.5.0"
-  val decline = "0.6.0"
-  val slickMigrationAPI = "0.4.0"
-  val spark = "2.2.0"
-  val sup = "0.2.0"
-}
