+++ app-backend/project/plugins.sbt
@@ -10,8 +10,6 @@ addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.14.6")
 
 addSbtPlugin("io.spray" % "sbt-revolver" % "0.9.1")
 
-addSbtPlugin("com.sksamuel.scapegoat" %% "sbt-scapegoat" % "1.0.9")
-
 addSbtPlugin("net.virtual-void" % "sbt-dependency-graph" % "0.9.0")
 
 addSbtPlugin("com.typesafe.sbt" % "sbt-git" % "1.0.0")
@@ -22,6 +20,8 @@ addSbtPlugin("com.github.gseitz" % "sbt-release" % "1.0.10")
 
 addSbtPlugin("io.get-coursier" % "sbt-coursier" % "1.0.3")
 
-addSbtPlugin("com.github.cb372" % "sbt-explicit-dependencies" % "0.2.8")
+addSbtPlugin("com.github.cb372" % "sbt-explicit-dependencies" % "0.2.9")
 
 addSbtPlugin("ch.epfl.scala" % "sbt-scalafix" % "0.9.1")
+
+addSbtPlugin("com.sksamuel.scapegoat" %% "sbt-scapegoat" % "1.0.9")
